Add RemoveStakesOnLastConfirmed function to remove staked validators on last confirmed assertion
This PR adds a new function RemoveStakesOnLastConfirmed to the Service struct in order to remove all staked validators on the last confirmed assertion. This function makes use of the GetLastConfirmedAssertion function from the L1 client to get the last confirmed assertion ID. It then filters all StakerStaked events for that assertion ID and removes the stakes for each staker if they staked on that assertion ID.

This function can be useful when we need to flush the validators from the system who have staked on a prior existing assertion which is already confirmed. This is required so that we can call confirm on the latest assertion.

This PR also includes the necessary error handling for each step of the process, so that any errors are properly handled and returned to the caller.